requirements.txt
python-telegram-bot[job-queue]
python-dotenv
requests
opencv-python
RPi.GPIO
